Introduce RuleCarrier for carrying context around rule inspection
Move ASTContext and ViolationSet from RuleBase to RuleCarrier, so that
RuleBase becomes a pure interface but takes a RuleCarrier. RuleCarrier can 
be upgraded, but RuleBase interface will keep the same. 

In addition, methods like adding violations to violation set now are extracted 
to RuleCarrier, which makes more sense for logic separation.