Fix DiagnosticConsumer memory leak
The DiagnosticConsumer gets allocated with new but is never deleted. The
engine is explicitly told to not take ownership of the consumer. This
would make sense if the consumer is later freed by oclint, but this
never happens. Instead we should just let the engine take ownership of
the client so that it gets deleted alongside the enngine.