Fix segfault in NPathComplexityMetric
The nPath method for SwitchStmt casts the body to CompoundStmt. A switch
statement can also contain just a single case like this:
switch(arg)
   case3: return 2;

Here the body is actually just a single SwitchStmt. We have to check the
actual type of the body before casting it to CompoundStmt and
potentially triggering UB.