resolve path correctly for locations indentified as macros
LLVM seems to identify some lines as macros, and doesn't compute the
filename directly for them. The diagnostic printer now tries to expand
the macro when no filename can be resolved in the first try.

This approach was chosen mostly to be backwards compatible with cases
where for some reason, the first attempt works for macro locations.