Fix segfault on methods without CompoundStmt
It is legal in c++ to leave out the curly braces of function and just
have a single statement like a try-catch or a while loop as the function
body. If we encounter such a method in applyDecl, the dyn_cast will fail
because it is not a CompoundStmt and we get a segfault on a nullptr.

Thankfully, the parent class Stmt also has the getSourceRange() method
that works just as expected from what I have tested so far.