bundle: fix cpp header install dir
If the llvm-root option is passed to bundle, the script attempts to
copy the cpp headers from an incorrect location (namely, where oclint
would build them itself).

This commit corrects the path to the cpp headers if the llvm-root
option is passed to bundle so that bundle copies the required headers
from the correct location within the existing LLVM install prefixed by
args.llvm_root.