Don't overinitialize the CompilerInstance
While debugging a problem in which clang would compile a given
file while oclint would report duplicate symbols, I traced the
problem to OCLint not honoring the -ivfsoverlay flag, thus finding
the header in two different paths and defeating the single-include
guard added for Objective-C.

This commit removes the manual initialization of the FileManager
and the SourceManager. They are automatically initialized properly
by the FrontEndAction's BeginSourceFile if they haven't been
initialized.

Letting BeginSourceFile initialize the FileManager seems safer than
copying the logic of how Clang initializes it.

The end result with this commit is that the -ivfsoverlay flag is
honoured by OCLint.